+++ lib/src/widgets/delegate.dart
@@ -19,12 +19,32 @@ typedef MentionBlockBuilder = Widget Function(
 
 typedef CustomStyleBuilder = TextStyle Function(Attribute attribute);
 
+/// Delegate interface for the [EditorTextSelectionGestureDetectorBuilder].
+///
+/// The interface is usually implemented by textfield implementations wrapping
+/// [EditableText], that use a [EditorTextSelectionGestureDetectorBuilder]
+/// to build a [EditorTextSelectionGestureDetector] for their [EditableText].
+/// The delegate provides the builder with information about the current state
+/// of the textfield.
+/// Based on these information, the builder adds the correct gesture handlers
+/// to the gesture detector.
+///
+/// See also:
+///
+///  * [TextField], which implements this delegate for the Material textfield.
+///  * [CupertinoTextField], which implements this delegate for the Cupertino
+///    textfield.
 abstract class EditorTextSelectionGestureDetectorBuilderDelegate {
-  GlobalKey<EditorState> getEditableTextKey();
+  /// [GlobalKey] to the [EditableText] for which the
+  /// [EditorTextSelectionGestureDetectorBuilder] will build
+  /// a [EditorTextSelectionGestureDetector].
+  GlobalKey<EditorState> get editableTextKey;
 
-  bool getForcePressEnabled();
+  /// Whether the textfield should respond to force presses.
+  bool get forcePressEnabled;
 
-  bool getSelectionEnabled();
+  /// Whether the user may select text in the textfield.
+  bool get selectionEnabled;
 }
 
 /// Builds a [EditorTextSelectionGestureDetector] to wrap an [EditableText].
@@ -69,16 +89,12 @@ class EditorTextSelectionGestureDetectorBuilder {
   /// The [State] of the [EditableText] for which the builder will provide a
   /// [EditorTextSelectionGestureDetector].
   @protected
-  EditorState? getEditor() {
-    return delegate.getEditableTextKey().currentState;
-  }
+  EditorState? get editor => delegate.editableTextKey.currentState;
 
   /// The [RenderObject] of the [EditableText] for which the builder will
   /// provide a [EditorTextSelectionGestureDetector].
   @protected
-  RenderEditor? getRenderEditor() {
-    return getEditor()!.getRenderEditor();
-  }
+  RenderEditor? get renderEditor => editor?.renderEditor;
 
   /// Handler for [EditorTextSelectionGestureDetector.onTapDown].
   ///
@@ -92,7 +108,7 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onTapDown(TapDownDetails details) {
-    getRenderEditor()!.handleTapDown(details);
+    renderEditor!.handleTapDown(details);
     // The selection overlay should only be shown when the user is interacting
     // through a touch screen (via either a finger or a stylus).
     // A mouse shouldn't trigger the selection overlay.
@@ -116,10 +132,10 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onForcePressStart(ForcePressDetails details) {
-    assert(delegate.getForcePressEnabled());
+    assert(delegate.forcePressEnabled);
     shouldShowSelectionToolbar = true;
-    if (delegate.getSelectionEnabled()) {
-      getRenderEditor()!.selectWordsInRange(
+    if (delegate.selectionEnabled) {
+      renderEditor!.selectWordsInRange(
         details.globalPosition,
         null,
         SelectionChangedCause.forcePress,
@@ -140,14 +156,14 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onForcePressEnd(ForcePressDetails details) {
-    assert(delegate.getForcePressEnabled());
-    getRenderEditor()!.selectWordsInRange(
+    assert(delegate.forcePressEnabled);
+    renderEditor!.selectWordsInRange(
       details.globalPosition,
       null,
       SelectionChangedCause.forcePress,
     );
     if (shouldShowSelectionToolbar) {
-      getEditor()!.showToolbar();
+      editor!.showToolbar();
     }
   }
 
@@ -161,8 +177,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///    this callback.
   @protected
   void onSingleTapUp(TapUpDetails details) {
-    if (delegate.getSelectionEnabled()) {
-      getRenderEditor()!.selectWordEdge(SelectionChangedCause.tap);
+    if (delegate.selectionEnabled) {
+      renderEditor!.selectWordEdge(SelectionChangedCause.tap);
     }
   }
 
@@ -190,8 +206,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onSingleLongTapStart(LongPressStartDetails details) {
-    if (delegate.getSelectionEnabled()) {
-      getRenderEditor()!.selectPositionAt(
+    if (delegate.selectionEnabled) {
+      renderEditor!.selectPositionAt(
         from: details.globalPosition,
         cause: SelectionChangedCause.longPress,
       );
@@ -209,8 +225,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///    triggers this callback.
   @protected
   void onSingleLongTapMoveUpdate(LongPressMoveUpdateDetails details) {
-    if (delegate.getSelectionEnabled()) {
-      getRenderEditor()!.selectPositionAt(
+    if (delegate.selectionEnabled) {
+      renderEditor!.selectPositionAt(
         from: details.globalPosition,
         cause: SelectionChangedCause.longPress,
       );
@@ -228,7 +244,7 @@ class EditorTextSelectionGestureDetectorBuilder {
   @protected
   void onSingleLongTapEnd(LongPressEndDetails details) {
     if (shouldShowSelectionToolbar) {
-      getEditor()!.showToolbar();
+      editor!.showToolbar();
     }
   }
 
@@ -243,10 +259,10 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onDoubleTapDown(TapDownDetails details) {
-    if (delegate.getSelectionEnabled()) {
-      getRenderEditor()!.selectWord(SelectionChangedCause.tap);
+    if (delegate.selectionEnabled) {
+      renderEditor!.selectWord(SelectionChangedCause.tap);
       if (shouldShowSelectionToolbar) {
-        getEditor()!.showToolbar();
+        editor!.showToolbar();
       }
     }
   }
@@ -261,7 +277,7 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onDragSelectionStart(DragStartDetails details) {
-    getRenderEditor()!.handleDragStart(details);
+    renderEditor!.handleDragStart(details);
   }
 
   /// Handler for [EditorTextSelectionGestureDetector.onDragSelectionUpdate].
@@ -276,7 +292,7 @@ class EditorTextSelectionGestureDetectorBuilder {
   @protected
   void onDragSelectionUpdate(
       DragStartDetails startDetails, DragUpdateDetails updateDetails) {
-    getRenderEditor()!.extendSelection(updateDetails.globalPosition,
+    renderEditor!.extendSelection(updateDetails.globalPosition,
         cause: SelectionChangedCause.drag);
   }
 
@@ -290,7 +306,7 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback.
   @protected
   void onDragSelectionEnd(DragEndDetails details) {
-    getRenderEditor()!.handleDragEnd(details);
+    renderEditor!.handleDragEnd(details);
   }
 
   /// Returns a [EditorTextSelectionGestureDetector] configured with
@@ -302,9 +318,8 @@ class EditorTextSelectionGestureDetectorBuilder {
     return EditorTextSelectionGestureDetector(
       key: key,
       onTapDown: onTapDown,
-      onForcePressStart:
-          delegate.getForcePressEnabled() ? onForcePressStart : null,
-      onForcePressEnd: delegate.getForcePressEnabled() ? onForcePressEnd : null,
+      onForcePressStart: delegate.forcePressEnabled ? onForcePressStart : null,
+      onForcePressEnd: delegate.forcePressEnabled ? onForcePressEnd : null,
       onSingleTapUp: onSingleTapUp,
       onSingleTapCancel: onSingleTapCancel,
       onSingleLongTapStart: onSingleLongTapStart,